OAK-12162 : moved cachestats and cachestatssnapshot in same package as that of abstractcachestats
