+++ oak-segment-tar/src/main/java/org/apache/jackrabbit/oak/segment/WriterCacheManager.java
@@ -30,7 +30,7 @@ import java.util.function.Predicate;
 import java.util.function.Supplier;
 
 import org.apache.jackrabbit.oak.api.jmx.CacheStatsMBean;
-import org.apache.jackrabbit.oak.cache.api.CacheStatsSnapshot;
+import org.apache.jackrabbit.oak.cache.CacheStatsSnapshot;
 import org.apache.jackrabbit.oak.commons.collections.IteratorUtils;
 import org.apache.jackrabbit.oak.segment.file.PriorityCache;
 import org.apache.jackrabbit.oak.stats.StatisticsProvider;
@@ -354,17 +354,11 @@ public abstract class WriterCacheManager {
         private static <T> Supplier<CacheStatsSnapshot> accumulateRecordCacheStats(
                 final Iterable<RecordCache<T>> caches) {
             return () -> {
-                long hits = 0, misses = 0, loads = 0, failures = 0, loadTime = 0, evictions = 0;
+                CacheStatsSnapshot stats = new CacheStatsSnapshot(0, 0, 0, 0, 0, 0);
                 for (RecordCache<?> cache : caches) {
-                    CacheStatsSnapshot s = cache.getStats();
-                    hits += s.hitCount();
-                    misses += s.missCount();
-                    loads += s.loadSuccessCount();
-                    failures += s.loadFailureCount();
-                    loadTime += s.totalLoadTime();
-                    evictions += s.evictionCount();
+                    stats = stats.plus(cache.getStats());
                 }
-                return new CacheStatsSnapshot(hits, misses, loads, failures, loadTime, evictions);
+                return stats;
             };
         }
