+++ oak-store-document/src/main/java/org/apache/jackrabbit/oak/plugins/document/DocumentNodeStoreBuilder.java
@@ -38,7 +38,7 @@ import java.util.function.Supplier;
 import org.apache.jackrabbit.oak.cache.CacheLIRS;
 import org.apache.jackrabbit.oak.cache.api.Cache;
 import org.apache.jackrabbit.oak.cache.api.CacheBuilder;
-import org.apache.jackrabbit.oak.cache.api.CacheStats;
+import org.apache.jackrabbit.oak.cache.CacheStats;
 import org.apache.jackrabbit.oak.cache.api.Weigher;
 import org.apache.jackrabbit.oak.cache.AbstractCacheStats;
 import org.apache.jackrabbit.oak.cache.CacheValue;
@@ -984,8 +984,8 @@ public class DocumentNodeStoreBuilder<T extends DocumentNodeStoreBuilder<T>> {
         return buildCache(CacheType.CHILDREN, getChildrenCacheSize(), store, null);
     }
 
-    public org.apache.jackrabbit.oak.cache.api.Cache<CacheValue, StringValue> buildMemoryDiffCache() {
-        return org.apache.jackrabbit.oak.cache.api.CacheBuilder
+    public Cache<CacheValue, StringValue> buildMemoryDiffCache() {
+        return CacheBuilder
                 .<CacheValue, StringValue>newBuilder()
                 .maximumWeight(getMemoryDiffCacheSize())
                 .weigher((k, v) -> weigher.weigh(k, v))
@@ -993,8 +993,8 @@ public class DocumentNodeStoreBuilder<T extends DocumentNodeStoreBuilder<T>> {
                 .build();
     }
 
-    public org.apache.jackrabbit.oak.cache.api.Cache<RevisionsKey, LocalDiffCache.Diff> buildLocalDiffCache() {
-        return org.apache.jackrabbit.oak.cache.api.CacheBuilder
+    public Cache<RevisionsKey, LocalDiffCache.Diff> buildLocalDiffCache() {
+        return CacheBuilder
                 .<RevisionsKey, LocalDiffCache.Diff>newBuilder()
                 .maximumWeight(getLocalDiffCacheSize())
                 .weigher((k, v) -> weigher.weigh(k, v))
